+++ lib/ClustersBackend/PlacesBackend.php
@@ -71,6 +71,9 @@ class PlacesBackend extends Backend
         $count = $query->func()->count($query->createFunction('DISTINCT m.fileid'), 'count');
         $query->select('e.osm_id', 'e.name', $count)->from('memories_planet', 'e');
 
+        // WHERE these are not special clusters (e.g. timezone)
+        $query->where($query->expr()->gt('e.admin_level', $query->createNamedParameter(0)));
+
         // WHERE there are items with this osm_id
         $query->innerJoin('e', 'memories_places', 'mp', $query->expr()->eq('mp.osm_id', 'e.osm_id'));